requirements.txt
numpy>=1.23.5,<2.0.0
pandas>=1.5.3,<2.0.0
xgboost>=1.7.5,<2.0.0
optuna>=3.2.0,<4.0.0
yfinance>=0.2.18,<1.0.0
python-telegram-bot>=20.3,<21.0.0
colorama>=0.4.6,<1.0.0
ta>=0.10.2,<1.0.0
SQLAlchemy>=1.4.46,<2.0.0
PyYAML>=6.0.1,<7.0.0
scikit-learn>=1.3.0,<2.0.0
schedule>=1.2.0,<2.0.0
aiohttp>=3.8.5,<4.0.0